Added: YKEnterISR, YKExitISR, YKTickHandler
Also edited makefile to use Lab4c.c.
Have a very weird bug in YKTickHandler where incrementing YKTickNum causes compiler error, so I commented that out.
When running without that, we get an illegal opcode. It looks like the ip is out of range, which could indicate a problem with saving contexts.